backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets>=13.0,<15.0
python-dotenv==1.0.0
google-genai==0.2.0
python-multipart==0.0.6

